Use `http-proxy` to forward requests to _either_ the rendering server or the asset server. This obviates the need for the browser sending requests on different ports. Because of this, the `publicPath` can be just that: a path (since it now looks like it resides on the same server).
The `webpack-dev-server` was stripped out and replaced with the simpler `webpack-dev-middleware` since the server does not allow controlling the HMR `socket.io` path which is necessary to tweak with these changes. It's a little bit messy but it's where we're at for now. It also means the asset server behaves much more like a "real" http server in the same vein the rendering server does.

It also means that you can view your page "instantly" while the render server sets up; the proxy in front will simply wait while the render server boots up and then forward the request when it's ready.

Interestingly enough this should not be a breaking change since no options to the origin dev server were ever really exposed.