Only make `runtime` return data when running in server.
The only time the runtime information is actually useful is when something is being run under the dev server. In these cases the webpack config that is included by the dev server will invoke the call to runtime and the relevant data will be returned. In other cases it will simply be an empty array.

This is kind of hacky and maybe there's a better solution out there, but this will at least fix the issue where the runtime is included in a non-udev-server dev build causing it to crash.