Whole new dev-server implementation.
Pretty much a total rewrite, with a ton of great new features and architecture:

 * Parallelized builds for multi-webpack configurations.
 * Sick emoji logging.
 * Custom HMR runtime.
 * Hot-reloading of webpack configuration files.
 * Universal, detachable IPC.
 * Simple, dynamic proxying.
 * Host multiple applications at once.
 * Dump your webpack configuration for debugging.

The architecture is now such that:

 * There is a simple global "master" process that acts as a smart proxy.
 * This process watches for webpack configurations.
 * Each configuration gets spawned a compiler instance in a new process.
 * Each compiler is responsible for spawning the related app:
  * The built bundle for `node` types,
  * The asset server for `web` types.
 * Everything that needs a URL registers itself to the smart proxy.
 * Asset stats can now be monitored by anyone on the IPC net.

The whole HMR runtime has been rewritten to work with this architecture and now no longer touches webpack's rather nasty version. Because every part communicates over HTTP IPC, it's possible to run each child compiler in its own terminal and attach manually. Because the platform is agnostic towards individual configurations and can mount apps at arbitrary points, it's possible to run multiple webpack applications on this single platform.

In order for the asset server to advertise its stats information to other processes the IPC network can be used. There's now a built-in middleware called `dev-assets` which does exactly this, and attaches a `req.assets` property to every request.

The new packages used:

 * `http-middleware-metalab` for beautiful server/proxying/etc.
 * `chokidar` for file watching.
 * `purdy` for dumping the webpack configuration in pretty colors.
 * `memory-fs` for the asset server.