Hold connections open until `webpack` processes are done.
Previously if you had stuff in flight the dev server would display its status page. Sometimes this can be frustrating and you would rather just have the browser wait until the page is ready and then load that. This is now the case. Unfortunately this kind of messes with `404` handling - to do `404` handling properly we need either timeouts or to still allow for a window of "blankness" where we have to wait until the server registers pages which are meant to be proxied. Either way this seems like a good start.