Refactor everything again.
Hopefully this is close to the last one for a while. The biggest changes here are:

 * Switch to real protocol (GraphQL) instead of DIY websocket.
 * Pull out webpack-specific stuff into its own package.
 * Add a UI built on `Apollo` + `blessed`.

There are a bunch of smaller updates along the way too, however:

 * Update all the packages.
 * Allow handling extension types via `-r` argument.
 * Allow configuration and instantiation of server through `createServer`.
 * Allow HTTP add-ons akin to `webpack-serve` via `--add`.
 * Allow `publicPath` to contain URL values.
 * Fix intermittent proxy bug.
 * Switch to `jest` for testing.
 * Add more tests.
 * Log HTTP requests for you.

A lot of this isn't totally finished but it's good enough to get out the door and start tire kicking. There will be plenty of smaller PRs to fix bugs and issues that are already known.